create OAuthConsentServlet to control OAuth flow
